Add subdirectory 'common' under 'components'
