+++ front/components/common/Logo.js
@@ -1,3 +1,4 @@
+import React from "react";
 
 export default function Logo() {
   return (
